Keep Node files out of utils
