+++ src/bin/constants/cli.ts
@@ -1,0 +1,3 @@
+export const DESCRIPTION = "Generates one or more words|sentences|paragraphs";
+
+export const USAGE = "3 words [options]";
