+++ src/bin/constants/commands.ts
@@ -1,0 +1,5 @@
+export const COPY = {
+  DARWIN: "pbcopy",
+  LINUX: "xclip -selection clipboar",
+  WIN32: "clip",
+};
