+++ src/bin/constants/errors.ts
@@ -1,0 +1,2 @@
+export const CANNOT_DETERMINE_PLATFORM =
+  "Could not determine host operating system.";
