+++ src/bin/util/getVersion.ts
@@ -1,0 +1,9 @@
+import fs from "fs";
+import { version } from '../../../package.json'
+
+/**
+ * @returns  The package version.
+ */
+const getVersion = (): string => version
+
+export default getVersion;
