+++ src/bin/util/index.ts
@@ -1,0 +1,13 @@
+import copyToClipboard from './copyToClipboard'
+import getCopyCommand from './getCopyCommand'
+import getPlatform from './getPlatform'
+import getVersion from './getVersion'
+import isSupportedPlatform from './isSupportedPlatform'
+
+export {
+  copyToClipboard,
+  getCopyCommand,
+  getPlatform,
+  getVersion,
+  isSupportedPlatform,
+}
