--- src/constants/cli.ts
@@ -1,3 +1,0 @@
-export const DESCRIPTION = "Generates one or more words|sentences|paragraphs";
-
-export const USAGE = "3 words [options]";
