--- src/constants/commands.ts
@@ -1,5 +1,0 @@
-export const COPY = {
-  DARWIN: "pbcopy",
-  LINUX: "xclip -selection clipboar",
-  WIN32: "clip",
-};
