--- src/constants/errors.ts
@@ -1,2 +1,0 @@
-export const CANNOT_DETERMINE_PLATFORM =
-  "Could not determine host operating system.";
