--- src/util/getVersion.ts
@@ -1,9 +1,0 @@
-import fs from "fs";
-import { version } from '../../package.json'
-
-/**
- * @returns  The package version.
- */
-const getVersion = (): string => version
-
-export default getVersion;
