+++ src/util/index.ts
@@ -1,24 +1,14 @@
 import capitalize from "./capitalize";
-import copyToClipboard from "./copyToClipboard";
-import getCopyCommand from "./getCopyCommand";
-import getPlatform from "./getPlatform";
-import getVersion from "./getVersion";
 import isNode from "./isNode";
 import isReactNative from "./isReactNative";
-import isSupportedPlatform from "./isSupportedPlatform";
 import isWindows from "./isWindows";
 import makeArrayOfLength from "./makeArrayOfLength";
 import makeArrayOfStrings from "./makeArrayOfStrings";
 
 export {
   capitalize,
-  copyToClipboard,
-  getCopyCommand,
-  getPlatform,
-  getVersion,
   isNode,
   isReactNative,
-  isSupportedPlatform,
   isWindows,
   makeArrayOfLength,
   makeArrayOfStrings,
